change interface for invocation from .invokeLambdaFunction to Lambda.invoke fixes #4
