+++ .gitignore
@@ -32,3 +32,4 @@ node_modules
 # Optional REPL history
 .node_repl_history
 src/
+.DS_Store
