+++ lib/index.js
@@ -1,58 +1,65 @@
 'use strict';
 
- // Context example
- // {
- //    "awsRequestId": "ebb337f8-dbe6-11e5-993d-8d772de156ef",
- //    "invokeid": "ebb337f8-dbe6-11e5-993d-8d772de156ef",
- //    "logGroupName": "/aws/lambda/aws-canary-lambda",
- //    "logStreamName": "2016/02/25/[$LATEST]af43675df7bd4d128d14ea3669d45635",
- //    "functionName": "aws-canary-lambda",
- //    "memoryLimitInMB": "128",
- //    "functionVersion": "$LATEST",
- //    "invokedFunctionArn": "arn:aws:lambda:eu-west-1:447022989235:function:aws-canary-lambda"
- //  }
+var a = {}; // re-useable Object
 
-class AwsHelper {
-  constructor (aws, context, config) {
-    this._aws = aws;
-    this._context = context || {};
-    this._config = config || {};
-    this._extractParams(this._context.invokedFunctionArn || '');
+function extractParams (invokedFunctionArn) {
+  // console.log(' >> ', invokedFunctionArn);
+  var invokedFunctionArnArray = invokedFunctionArn.split(':');
+  // console.log(' >> invokedFunctionArnArray.length: ', invokedFunctionArnArray.length);
+  // no label provided : e.g. "invokedFunctionArn": "arn:aws:lambda:eu-west-1:447022989235:function:aws-canary-lambda"
+  if (invokedFunctionArnArray.length === 7) {
+    a.region = invokedFunctionArnArray[3];
+    a.account = invokedFunctionArnArray[4];
+    a.version = '$LATEST';
+    return a;
+  }
+  // label provided : e.g. "invokedFunctionArn": "arn:aws:lambda:eu-west-1:447022989235:function:aws-canary-lambda:ci"
+  if (invokedFunctionArnArray.length === 8) {
+    a.region = invokedFunctionArnArray[3];
+    a.account = invokedFunctionArnArray[4];
+    a.version = a.env = invokedFunctionArnArray[7];
+    return a;
   }
 
-  _extractParams (invokedFunctionArn) {
-    let invokedFunctionArnArray = invokedFunctionArn.split(':');
-    // no label provided : e.g. "invokedFunctionArn": "arn:aws:lambda:eu-west-1:447022989235:function:aws-canary-lambda"
-    if (invokedFunctionArnArray === 7) {
-      this.region = invokedFunctionArnArray[3];
-      this.accountId = invokedFunctionArnArray[4];
-      this.version = '$LATEST';
-      return;
-    }
-    // label provided : e.g. "invokedFunctionArn": "arn:aws:lambda:eu-west-1:447022989235:function:aws-canary-lambda:ci"
-    if (invokedFunctionArnArray === 8) {
-      this.region = invokedFunctionArnArray[3];
-      this.accountId = invokedFunctionArnArray[4];
-      this.version = '$LATEST';
-      return;
-    }
-
-    throw new Error('Unexpected invokedFunctionArn format: ' + invokedFunctionArn);
-  }
-
-  invokeLambdaFunction (lambdaFunctionMame, payload, cb) {
-    if (!lambdaFunctionMame) throw new Error('lambdaFunctionMame is required');
-    let params = {
-      FunctionName: this.accountId + ':' + lambdaFunctionMame,
-      InvocationType: this._config.lambdaInvocationType || 'RequestResponse',
-      Payload: JSON.stringify(payload),
-      Qualifier: this.version,
-      LogType: this._config.lambdaInvocationType || 'None'
-    };
-    this._aws.ambda.invoke(params, cb);
-  }
+  throw new Error('Unexpected invokedFunctionArn format: ' + invokedFunctionArn);
 }
 
-export default {
-  AwsHelper
-};
+// initialisation method (Optional)
+function AwsHelper (aws, context, config) {
+  a._aws = aws;
+  a._context = context;
+  a._config = config || {};
+  extractParams(context.invokedFunctionArn);
+  return a;
+}
+
+function getEnvironment (context) {
+  if (!context || !context.invokedFunctionArn) {
+    return null;
+  }
+  var env = a.env || extractParams(context.invokedFunctionArn).env;
+  return env;
+}
+
+var AWS = require('aws-sdk');
+// http://docs.aws.amazon.com/AWSJavaScriptSDK/latest/AWS/Lambda.html#invoke-property
+var Lambda = {};
+
+Lambda.invoke = function (params, cb) {
+  if (!params || !params.FunctionName) {
+    throw new Error('params.FunctionName is required');
+  }
+  AWS.config.region = a.region;
+  var lambda = new AWS.Lambda();
+  var p = {
+    FunctionName: a.account + ':' + params.FunctionName,
+    InvocationType: a._config.lambdaInvocationType || 'RequestResponse',
+    Payload: JSON.stringify(params.Payload),
+    Qualifier: params.Qualifier || a.version,
+    LogType: a._config.lambdaInvocationType || 'None'
+  };
+  return lambda.invoke(p, cb);
+}
+module.exports = AwsHelper;
+module.exports.Lambda = Lambda;
+module.exports.getEnvironment = getEnvironment;
