--- mock.js
@@ -1,4 +1,0 @@
-var awsMock = require('aws-sdk-mock');
-console.log(awsMock);
-
-// console.log(JSON.stringify(awsMock, null, 2));
